+++ server/tests/models/MaterialTest.php
@@ -132,10 +132,11 @@ final class MaterialTest extends ModelTestCase
         $Material = $this->model::find(1);
         $result   = $Material->park;
         $this->assertEquals([
-            'id'           => 1,
-            'name'         => "default",
-            'total_items'  => 5,
-            'total_amount' => 19907.40,
+            'id' => 1,
+            'name' => "default",
+            'total_items' => 5,
+            'total_amount' => 101223.80,
+            'total_stock_quantity' => 83,
         ], $result);
     }