+++ server/tests/models/ParkTest.php
@@ -36,7 +36,7 @@ final class ParkTest extends ModelTestCase
     public function testGetTotalAmount(): void
     {
         $Park = $this->model::find(1);
-        $this->assertEquals(19907.40, $Park->total_amount);
+        $this->assertEquals(101223.80, $Park->total_amount);
     }
 
     public function testGetPerson(): void