+++ server/CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Update dependencies
 - Remove bills file storage, and always re-create PDFs on demand (#8).
 - Change bills numbers to be successive instead of using date of creation (#8).
+- Fix total replacement amount of parks material (#6).
 
 ## 0.8.2 (2020-07-02)